SignIn button state changes + refactoring
