+++ src/components/AccountButton.scss
@@ -1,28 +1,19 @@
-@mixin absolute-position {
+.account-button {
     position: absolute;
-    top: 1rem;
-    right: 1rem;
+    top: 0.5em;
+    left: 0.5em;
 }
 
-#sign-out {
-    @include absolute-position();
+.account-button-pic {
+    height: 1.5em;
+    height: 1.5em;
+    border-radius: 0.75em;
+    border-style: solid;
+    border-width: 1px;
+    margin-top: -0.45em;
+    margin-left: -0.1em;
 }
 
-#sign-in {
-    @include absolute-position();
-}
-
-.icon-button {
-    position: absolute;
-    top: 1rem;
-    left: 1rem;
-}
-
-.dialog-content {
-    text-align: center;
-}
-
-.dialog-card {
-    margin: auto;
-    width: 24rem;
+.account-button-text {
+    margin-left: 1em;
 }
