+++ src/components/SignInButton.scss
@@ -1,0 +1,28 @@
+@mixin absolute-position {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+}
+
+#sign-out {
+    @include absolute-position();
+}
+
+#sign-in {
+    @include absolute-position();
+}
+
+.icon-button {
+    position: absolute;
+    top: 1rem;
+    left: 1rem;
+}
+
+.dialog-content {
+    text-align: center;
+}
+
+.dialog-card {
+    margin: auto;
+    width: 24rem;
+}
